Add a basic IntroductionWatcher, just deletes messages if same user posts twice
Not a huge improvement, but we do seem to have a couple people who post once a day, or once a week. Introductions is a pretty slow channel so this won't really add any additional cost of running. Maybe slightly more memory, but we only hold user id of past 1000 messages, which is tiny.