Create FuzzTable.cs
A fuzz string is a string with alternative choices to pick randomly.  "A (ball|block)." becomes "A ball." or "A block."

A fuzz table is a list of alternate choices, which may be loaded from a file, and each line can optionally be treated as a fuzz string.

Useful for humanizing text without too much complexity or overhead.